Slim delayed PPM record snapshot.
Fold the base script into a single self-contained record train_gpt.py and drop the extra base_train_gpt.py dependency while keeping the fixed K=15 delayed PPM config.

Made-with: Cursor
